Finished a 'decent' paddle-ball bounce system
Now the paddle detects in which percentage of the half of the paddle the ball has collided with, and reflects the ball giving it that percentage to the velocity distributed in X and Y.

+++ Project/include/ball.h
@@ -24,8 +24,8 @@ public:
 
   Vector2D getPosition() const;
   Vector2D getVelocity() const;
-  uint32 getWidth() const;
-  uint32 getHeight() const;
+  float getWidth() const;
+  float getHeight() const;
   uint32 getNumBounces() const;
 
   void update(float dt);
